+++ src/components/auth/signup/agree/agreeForm.tsx
@@ -1,4 +1,4 @@
-import * as S from '../../../../styles/auth/signup/signup';
+import * as S from '../../../../styles/auth/signup/signup.style';
 import Check from '../../../../assets/images/auth/signup/check.png';
 import CheckClick from '../../../../assets/images/auth/signup/checkClick.png';
 import ListAgree from './list-agree';
