+++ src/components/auth/signup/profile/profileImage.tsx
@@ -1,4 +1,4 @@
-import * as S from '../../../../styles/auth/signup/signup';
+import * as S from '../../../../styles/auth/signup/signup.style';
 
 interface Props {
   preview?: string;
