+++ src/components/auth/test/onboarding/item-onboarding.tsx
@@ -1,5 +1,5 @@
 import type { OnboardingItem } from '../../../../types/auth/test/onboarding';
-import * as T from '../../../../styles/auth/test/test';
+import * as T from '../../../../styles/auth/test/test.style';
 import ListSelect from './list-select';
 
 interface ItemOnboardingProps {
