+++ src/components/auth/test/onboarding/item-select.tsx
@@ -1,4 +1,4 @@
-import * as T from '../../../../styles/auth/test/test';
+import * as T from '../../../../styles/auth/test/test.style';
 
 interface ItemSelectProps {
   id: number;
