+++ src/components/auth/test/onboarding/list-select.tsx
@@ -1,4 +1,4 @@
-import * as T from '../../../../styles/auth/test/test';
+import * as T from '../../../../styles/auth/test/test.style';
 import ItemSelect from './item-select';
 
 interface ListSelectProps {
