+++ src/components/auth/test/qna/item-answer.tsx
@@ -1,4 +1,4 @@
-import * as T from '../../../../styles/auth/test/test';
+import * as T from '../../../../styles/auth/test/test.style';
 
 interface ItemAnswerProps {
   answer: string;
