+++ src/styles/footer/footer.style.ts
